AOL adapter - store bid data in the bidmanager
Unlike the appnexus adapter, the aol adapter was storing its bid object
in a local object using the placement as key. This limits the bid
requests to only load ads for unique placements. For example, let's say
I have three 300x250 ads and I want to use the same placement id for
each 300x250 ad instance. Storing bid data with the placement as key
will only allow one of the three to be displayed because using the
placement key will result in bid data being overwritten. This commit
updates the aol adapter to work in a similar fashion to the appnexus
adapter, using the bidmanager to store bid data.